Use a new VolumeBar object each time
This allows the path settings take to be more easily traced, and allows
for cleaner code in my opinion. This required removing the hacky and
non-idiomatic -(void)destroyHUD method and replacing it with a
-(void)dealloc method that made proper use of the meomry management
system. I also reduced the function calls and memory allocations needed
by moving the check if an HUD was already showing to Tweak.xm, requiring
the execution of a completion block when the HUD finishes animating out.

TODO: make more functional (reduce state), refactor methods inside
VolumeBar.xm, fix tons of bugs