Fix preferences requiring a respring to apply
This was caused by the `UIColor *color` in Tweak.xm not being retained
for the entire length of time the device was active. This left a
dangling pointer to some random object in memory which was then set as
the background color for mainView.

This commit also removes the last vestiges of respring button logic from
the prefs bundle.