Hide the status bar when volume bar is showing
This required the addition of a subtweak to hook into applications,
because iOS handles the status bar in a really really weird way. Also
added is the ability to receive current orientation from apps, to be
used later to init the volume bar. The same IPC system used to do this
will be used to make the bar orient correctly in all apps.