Fix ringer slider bugs
Ringer slider previous was a UISlider created by the tweak, with custom
notification handling and a ton of bugy stuff to change the volume and
update it. I discovered the method on MPVolumeSlider
-(void)setVolumeAudioCategory:(BOOL)arg1 which allows me to use only the
MPVolumeView for both the Audio/Video and Ringtone volumes. This cleans
up the code immensely and fixes _all_ of the bugs with the ringer
slider.

This was a quick commit so expect some regressions.