Adjust label and slide handle colors
The color of the label and slide handle now adjust based on the
foreground color of the statusbar, i.e. whether the statusbar text is
more black or more white. This only adjusts at the start, so if the text
of the current statusbar changes halfway through nothing will change.
This may be functionality I want to implement later.