Prevent crashing Springboard when quickly pressed
More info in issue #9, this commit closes that. When the volume buttons
were quickly pressed, the hook is still waiting on the async call to the
app but the second time calls resetTimer on the VolumeBar object.
resetTimer was incorrectly written, with one crucial part not inside the
check to make sure it was safe.